Replaced react-use-websocket with socket.io
