+++ portal/package.json
@@ -33,7 +33,7 @@
         "react-dom": "^18.2.0",
         "react-router-dom": "^6.22.0",
         "react-scripts": "^5.0.1",
-        "react-use-websocket": "^4.7.0",
+        "socket.io-client": "^4.7.4",
         "styled-components": "^6.1.8",
         "three": "^0.161.0",
         "typescript": "^5.3.3",
