+++ portal/src/components/Dashboard.tsx
@@ -1,18 +1,12 @@
 import './Dashboard.css';
-import {Action} from '../App';
 import Server from './Server';
 import {Button, Col, Container, Row} from 'react-bootstrap';
 import {useState} from 'react';
 import DashboardEditor from './DashboardEditor';
 import {useServers} from '../api/UseServers';
 
-export interface DashboardProps {
-    action: Action;
-}
-
-function Dashboard(props: DashboardProps) {
+function Dashboard() {
     const [isInEditMode, setEditModeEnabled] = useState(false);
-    const {action} = props;
     const {data: servers} = useServers();
 
     // Weird hack to fix issues with @react-three/drei
@@ -22,13 +16,13 @@ function Dashboard(props: DashboardProps) {
     if (servers == null) return null;
 
     return isInEditMode ? (
-        <DashboardEditor servers={servers} action={action} closeEditMode={() => setEditModeEnabled(false)} />
+        <DashboardEditor servers={servers} closeEditMode={() => setEditModeEnabled(false)} />
     ) : (
         <Container fluid>
             {Object.values(servers).map((server) => (
                 <Row key={server.id} className='mt-3'>
                     <Col>
-                        <Server server={server} action={action} />
+                        <Server server={server} />
                     </Col>
                 </Row>
             ))}
