+++ portal/src/components/Server.tsx
@@ -4,15 +4,16 @@ import {useNavigate} from 'react-router-dom';
 import {CSSProperties} from 'react';
 import FuelInfo from './FuelInfo';
 import {Server as APIServer} from '../api/UseServers';
+import {useWebSocket} from '../api/UseWebSocket';
 
 export interface ServerProps {
     server: APIServer;
-    action: Action;
 }
 
 function Server(props: ServerProps) {
+    const {server} = props;
     const navigate = useNavigate();
-    const {server, action} = props;
+    const {action} = useWebSocket();
 
     return (
         <Accordion defaultActiveKey='0'>
