+++ portal/src/components/turtle/Turtle.tsx
@@ -1,15 +1,9 @@
 import {useParams} from 'react-router-dom';
-import {Action} from '../../App';
 import {useTurtle} from '../../api/UseTurtle';
 import HUD from './HUD';
 import Turtle3DMap from './Turtle3DMap';
 
-export interface TurtleProps {
-    action: Action;
-}
-
-function Turtle(props: TurtleProps) {
-    const {action} = props;
+function Turtle() {
     const {serverId, id} = useParams() as {serverId: string; id: string};
     const {data: turtle} = useTurtle(serverId, id);
 
@@ -17,8 +11,8 @@ function Turtle(props: TurtleProps) {
 
     return (
         <>
-            <Turtle3DMap action={action} />
-            <HUD action={action} />
+            <Turtle3DMap />
+            <HUD />
         </>
     );
 }
