+++ portal/src/index.tsx
@@ -5,13 +5,16 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import QueryProvider from './api/QueryProvider';
+import {WebSocketProvider} from './api/UseWebSocket';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
     <React.StrictMode>
         <BrowserRouter>
             <QueryProvider>
-                <App />
+                <WebSocketProvider>
+                    <App />
+                </WebSocketProvider>
             </QueryProvider>
         </BrowserRouter>
     </React.StrictMode>
