+++ server/package.json
@@ -13,11 +13,11 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@fastify/cors": "^9.0.0",
-    "@fastify/websocket": "^8.3.1",
+    "@fastify/cors": "^9.0.1",
     "better-sqlite3": "^9.3.0",
     "fastify": "^4.25.2",
     "fs-extra": "^11.2.0",
+    "socket.io": "^4.7.4",
     "uuid": "^9.0.1",
     "websocket": "^1.0.34",
     "winston": "^3.11.0",
